requirements.txt
selenium==4.12.0
webdriver-manager==4.0.0
python-dotenv==1.0.0
openai==1.2.0
pydantic==2.3.0
python-multipart==0.0.6
requests==2.31.0